Add $app->cache to talk to Memcached
If memcached-port is set in app.conf, we try to load a memcached client
module and provide a Cache::Memcached-compatible object to use. We die
if we can't find a client module, but not being able to connect to the
server is not considered an error—the cache object will operate as if
every get is a miss.

If an application whose code uses $app->cache is run without setting
memcached-port in app.conf, it will die horribly. I don't care.