Add a parameter validator
It takes a hash of column definitions and a hash of column values, and
returns a hash of column names and values if all is well, or a hash of
column names and corresponding errors (e.g. required value is missing,
or value is invalid). Callers can supply their own error messages and
regexps or callbacks for validation.