Testing the previous change means being able to find Wigeon::User
G::Table::rowclass now gets a default implementation that looks first in
the application's namespace and then in Gadwall's namespace for a module
whose name corresponds to the singular form of the table name. So we use
Wigeon::User instead of Gadwall::User, see the new roles, and the test
works as expected.

Subclasses that redefine rowclass are expected to return a full name or
do their own module loading magic.