Make _shadow_controllers work with MOJO_RELOAD
If MOJO_RELOAD was set (daemon --reload) Mojo::Loader->load($class)
would not check to see if $class->can('new') before require-ing the
class. That broke, of course, because our shadow controllers do not
correspond to files on disk.

So we pile another hack upon the first one: add an @INC hook that
returns just "1;" for the "text" of the shadowy modules.