Use Data::Entropy+Crypt::Rijndael as a cryptographically secure PRNG
We don't want to use /dev/urandom's lower-quality PRNG, and we can't
afford to use /dev/random's higher-quality PRNG too often (because of
limited entropy availability). So we do the next best thing, which is
seeding an AES-CTR generator once at startup from /dev/random. There
may be a delay at the start, but that's all.

(In development, we can use /dev/urandom instead.)