Redirect to https before serving a login page
This depends on the frontend web server setting X-Forwarded-Protocol and
X-Forwarded-Host properly (as described in docs/gadwall/apache). If it's
not configured correctly, we might do something stupid (e.g. infinite
redirect loop). But that's better than leaking plaintext cookies.

Speaking of which, we should set ->secure(1) on the session cookie too.

It may be necessary to put the canonical URL into the configuration
rather than using X-Forwarded-Host, but we can worry about it later.