Make _shadow_controllers less brittle
The code used to break on 5.10.0 (and it was easier to rewrite the
function than upgrade Debian's perl packages). Now we push a custom @INC
sub for each shadowed module which returns exactly what one would write
in a real file to define a class inheriting from a Gadwall:: controller.
Nothing more (preloading, @ISA setting) is needed.