Store the __source URL in the session cookie, not the form
The __source parameter was an XSS vector, and there's no way to validate
it (making it a route name and asking "does a route named $x exist?" is
"<sri> untested, unsupported, you're on your own there" territory).

What we need is integrity protection, so we depend on the signed cookie
to give us that for the duration of the login process.

To my surprise, this didn't break tests, because I happened to be
passing as __source the same URL I had just ->get_ok'ed earlier.