Allow setting canonical-host and canonical-https-port
We use them (if set) to generate redirects using the canonical site name
(rather than e.g., 127.0.0.1:3000) instead of depending on mod_proxy to
rewrite our Location: headers. This is necessary if we use non-standard
http or https ports and can't expect redirects to plain "http://a" and
"https://a" to work.