Move all request-base rewriting into the before_dispatch hook
If X-Forwarded-Protocol is available, we use it (as before, but we use
only the last value). If the canonical name and ports are specified, we
use them too. Otherwise we use X-Forwarded-Host if available. If none of
them are available, we'll do nothing (using Host as before) and blindly
hope someone is rewriting our redirects.