Add a first-draft "forgot password" implementation
This chunk of code contains many pieces: (a) a new Controller that
generates signed "gate pass" links and provides a bridge to verify them,
(b) new actions in G::Users that send the user such a link, and process
the password reset request, (c) a G::Controller::canonical_url function,
(d) some configuration changes, (e) a few small bugfixes.

This works, but is hard to test automatically. It needs some tweaking,
mostly in the area of template/output generation. The code is sound.