Write out the shadow controllers in full for Wigeon
Whatever _shadow_controllers does breaks (as in SEGV) Mojo::Loader under
perl -d as soon as it tries to load one of the shadowed modules. I don't
have the inclination to try to debug the problem, so I'll just write out
the files for the test suite.

One hopes the bug won't bite in normal applications.