Make json_fragment wrap the response in a textarea on demand
This is a concession to jquery.form.js, which must use an iframe to
upload files from an AJAX form, and (for various horrible reasons)
has to cope with browsers not dealing with plain JSON responses.

(I used to override json_fragment in classes which needed this, but
that wasn't fine-grained enough, and made me have to split things up
unnecessarily. Setting the format in the stash is easier.)