Add a prototype queue processor
This consists of a daemon that LISTENs for notifications and scans the
queue table, handing each row off to a Gadwall::Dequeue::* handler to
process. There's not much error handling, other than "don't fall over".
It's simple-minded but robust, to match the queue table.

There's only one handler (::Mail), and it's unimplemented.