A very simple implementation of Dequeue::Mail
This assumes that, since the job has been submitted using enqueue_mail
and passed _prepare_mail, it is still fine to be passed on to mail().
That's good enough for now.

In any case, we don't know if we'll continue to use MIME::Lite in the
future. The module is not maintained or recommended (but it sure is a
convenient alternative to the current recommendations!).