Use a citext column for users.email
This is a big and messy change.

First, it means we'll treat email addresses as being case-insensitive.
In theory, localparts can be case-sensitive, so this is bad and wrong.
In practice, however, it's exactly what we want.

Second, the upgrade has two problems: we can't create the extension in
the upgrade script (only postgres can run it), and we can't just rerun
the 000-extensions.sql script (without 9.3's "if not exists") either.
Also, the "alter table …" statement can fail because the database may
already contain email addresses that differ only by case. Both these
problems must be dealt with by hand.