Always use /dev/urandom to seed our AES-CTR generator
This was motivated by http://www.2uo.de/myths-about-urandom/, which
explains that the scaremongering about /dev/urandom in the random(4)
manpage on Linux is unjustified and incorrect. Once the entropy pool
has been initialised at startup, both devices are a source of "good"
secure randomness (in fact, the _same_ randomness).

Given this revelation, it would be feasible to do away with our AES-CTR
generator altogether, and just read from /dev/urandom whenever we want
any randomness. But I'm not quite /that/ motivated yet.

We assume that we're running on Linux (or a system whose /dev/urandom
behaves in the same way), and take it for granted that the system has
been properly seeded at startup.

Finally, the annoying pause on startup (especially infuriating with
commands like "hypnotoad --stop ./app") is a thing of the past.