Make dequeued somewhat more "pg_ctl restart" friendly
If we wake up and find that the database connection has gone away, we
sleep for a while and try to connect again.

This isn't the whole story, however. There are two other calls to
$dbh->pg_notifies that, although much less likely, may fail. But
this commit fixes the most likely problem.

Also some minor cleanups: don't set MOJO_MODE=production (already the
default), avoid $dbh->func for the more direct accessors, etc.