Use $app->dbh instead of DBI->connect in tests
Random musing: the only reason I've had to call $app->dbh so far is to
connect as the database owner, and here are two more such cases. Yet the
application does not know the name of the database owner. Is that a bug?
Should the owner always be called ${db_user}_admin?