Make the whole dequeued main loop restart-friendly
Introduce a get_notify function that knows how to (a) call pg_notifies
in a loop until it gets a result to return, or (b) ditch the connection
and try to establish another one after a delay before going back to (a).

Simplify main to just call get_notify in an infinite loop.