Make the dequeued queue name configurable
The earlier hardcoded "queue" is just too generic, and I'm growing wary
of namespace conflicts. In theory, using the application name makes it
possible to run multiple dequeued processes (not that I need to). With
some configuration, it's also possible to make multiple applications
share a single dequeued (not that I need to do that either).